Move enum file
Change marker action create to display and initialize to create
